Fix pandas data round trip through impyla columnar HS2 fetching
