+++ ibis/sql/tests/test_compiler.py
@@ -151,10 +151,14 @@ WHERE `c` > 0"""
         query = ast.queries[0]
 
         sql_query = query.compile()
-        expected = """SELECT `g`, sum(`f`) AS `total`
-FROM alltypes
-WHERE `c` > 0
-GROUP BY 1"""
+        expected = """\
+SELECT `g`
+FROM (
+  SELECT `g`, sum(`f`) AS `total`
+  FROM alltypes
+  WHERE `c` > 0
+  GROUP BY 1
+) t0"""
 
         assert sql_query == expected
